backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
cryptography
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
python-multipart

# Planned for later project steps:
boto3
pytest
httpx
